Added react routing
Added react routing in the App.js component. Also created a Home and NotFound components which exist in the components directory. Lastly, Moved the LoggingInAndOut components to the components folder.
