+++ src/components/LoggingInAndOut.js
@@ -1,5 +1,5 @@
-import {firebase, auth} from './fire.js';
-import * as React from 'react';
+import {firebase, auth} from '../fire.js';
+import React from 'react';
 import { Button } from '@mui/material';
 
 
@@ -11,17 +11,14 @@ export function SignIn() {
     }
   
     return (
-      <>
-      
-        <Button variant="contained" className="sign-in" onClick={signInWithGoogle}>Sign in with Google</Button>
-      </>
+        <Button variant="contained" onClick={signInWithGoogle}>Sign in with Google</Button>
     )
   
   }
 
-export  function SignOut() {
+export function SignOut() {
     return auth.currentUser && (
       <button className="sign-out" onClick={() => auth.signOut()}>Sign Out</button>
     )
-  }
+}
