+++ src/components/NotFound.js
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const NotFound = () => {
+    return (
+        <div>
+            <strong>Page does not exist</strong>
+        </div>
+    )
+}
+
+export default NotFound;
